Crate db_sync_sqlx

Macros

sqlx_implementations_for_wrapper

Generates sqlx implementations for an unsigned wrapper of types that are signed. We expect that values will have always 0 as the most significant bit. For example TxIndex is in range of [0, 2^15-1], it will be u16 in domain, but it requires encoding and decoding like i16. See txindex, word31 and word63 types in db-sync schema definition.

Structs

Address

Asset

AssetName

BlockNumber

EpochNumber

EpochNumberRow

MainchainEpochNonce

PolicyId

SlotNumber

StakeDelegation

CREATE DOMAIN “lovelace” AS numeric(20,0) CONSTRAINT flyway_needs_this CHECK (VALUE >= 0::numeric AND VALUE <= ‘18446744073709551615’::numeric);

TxIndex

CREATE DOMAIN txindex AS smallint CONSTRAINT txindex_check CHECK ((VALUE >= 0));

TxIndexInBlock

TxValue

CREATE DOMAIN int65type AS numeric (20, 0) CHECK (VALUE >= -18446744073709551615 AND VALUE <= 18446744073709551615);